move common.css to the common directory
