+++ wiseService/vueapp/src/main.js
@@ -7,7 +7,7 @@ import store from './store';
 import { BootstrapVue, IconsPlugin } from 'bootstrap-vue';
 import 'bootstrap/dist/css/bootstrap.css';
 import 'bootstrap-vue/dist/bootstrap-vue.css';
-import '../../../common.css';
+import '../../../common/common.css';
 
 Vue.config.productionTip = false;
